refactor(viewer): extract the map camera into useMapCamera
ViewerMap.vue is ~2.4k lines of imperative canvas code. Pull the automatic
camera — auto-zoom (frame all players / a focused subset / a tracked world
point) and single-player follow — into `useMapCamera`, following the
`useGrenadeRendering` injection pattern already used in this component.

The composable owns its requestAnimationFrame loops, the two watchers that
start/stop them, and `followZoom`; it drives the shared view transform through
a small accessor interface (`zoom` ref + get/setPan + getL/getView), so the
pan-drag handlers and draw loop that stay in the component are untouched. Logic
is moved verbatim; only pan access and prop reads are rewired through getters.

ViewerMap: 2488 -> 2369 lines. type-check and production build pass.
